pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-esi-status"
description = "A simple Alliance Auth app to check the status of ESI and its endpoints"
readme = "README.md"
keywords = [
    "allianceauth",
    "esi",
    "esistatus",
    "eveonline",
]
license = { file = "LICENSE" }
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.6,<5",
    "allianceauth-app-utils>=1.12",
]
optional-dependencies.tests-allianceauth-latest = [
    "coverage",
    "django-webtest",
]
urls.Changelog = "https://github.com/ppfeufer/aa-esi-status/blob/master/CHANGELOG.md"
urls.Documentation = "https://github.com/ppfeufer/aa-esi-status/blob/master/README.md"
urls.Donations = "https://ko-fi.com/ppfeufer"
urls.Homepage = "https://github.com/ppfeufer/aa-esi-status"
urls.Source = "https://github.com/ppfeufer/aa-esi-status.git"
urls.Tracker = "https://github.com/ppfeufer/aa-esi-status/issues"

[tool.hatch.build]
include = [
    "/esistatus",
]

[tool.hatch.version]
path = "esistatus/__init__.py"
